backend/pyproject.toml
[project]
name = "clinical-insight-backend"
version = "0.1.0"
description = "Backend for Clinical Trial Insight Platform"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "python-multipart>=0.0.9",
    "google-generativeai>=0.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
